product_service/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
grpcio==1.74.0
grpcio-tools==1.74.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
protobuf==6.31.1
pyodbc==5.2.0
python-dotenv==1.1.1
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
